Add part references in GroupDefeatReward
Point.bt has a comment that these are part indices.
They have been indices in randomizer for a while now.

This may be tricky to fix if any of them have gotten offset in people's maps.
I made a basic dump of vanilla ones, but something automated may be needed:
<https://gist.github.com/gracenotes/592b59b35fc98180eb29022619079ac4>

But let's fix this now before future issues arise.