Update The SpEffect Row Names In Accordance To Player HKS
Updated names for SpEffects found in the HKS file of the player (from a version pre 1.07, so some SpEffects like the BHS and new crouch vs dodge attacks may be missing), with a short description of their usages. Modified SpEffects can be found with prefix (HKS). Some SpEffects are marked as "Unknown Function" with a description of what I think they do.