Add Arbitrary Rotation X/Y action, Reset Rotation action
Adds an action to rotate an object around the X and Y axis by a fixed amount with the press of a keybind (J for X axis, K for Y axis). The fixed amount for both can be changed in the Map Settings menu.

Adds an action to reset the current rotation of an object to 0, 0, 0 with the press of a keybind (L).