Fixing incorrect mapped values and adding back unambiguously correct manual values
1. All padding is zero
2. No duplicate field names
3. All fields aligned correctly (in some cases reverted 95% wrong params with no new info - see below)
4. All unknown offsets labeled correctly
5. Incorrect NT fields removed
6. Restored as many unambiguously correct unmapped types as possible

Notes on params:
AssetEnvironmentGeometryParam - This changed a fair bit vs ER, and some types are blatantly wrong later on. Just leave the blanket rewrite for now, since the fields are also wrong, but they're not more wrong than before.
AssetModelSfxParam - All manual fields are correct, there's no reason to change them
CameraFadeParam - This doesn't seem to be part of gameparam, left it alone
CylinderParam - Most fields now have impossible alignment at the moment, and the auto defs don't have any new info (they just shorten types). We should roll back to the previous version for now, as this is blatantly incorrect for all fields except for the one the mapper has info on.
DamageLevelConvParam - Also bad alignment
Depthline - Mapper is wrong - this is two floats. It may be using SIMD instructions to read it.
EFID_common - I haven't dumped this fully, but looks the same as Elden Ring
EquipParamGoods - The Elden Ring specific bitfields are pretty screwed up now, but I won't try to fix them here
EquipParamProtector - Bad alignment. I've tried to manually fixed these
GameSystemParam - Rather screwed up, most of the fields here are floats which have been rewritten to ints. But the initial type offsets were wrong so this can be redone later.
GrassTypeParam - Padding was invalid
HitEffectSfxConceptParam - Padding was invalid
LoadBalancerParam - Padding was invalid
SeMaterialConvParam - These types look wrong, but it's at least not padding.
ShopLineupParam - Mapper is wrong. 0x16 is not an int, it's a short.
TentativePlayerParam - There is a comment here. Renamed the field so as not to lose it
ThrustersParam - Dual s8 could also be an s16, and u16 could also be two u8s. They should definitely not be ints, though.
WorldMapLegacyConvParam - Major issue was offset by 1 because of ERNT fields. Removing these worked great. Not sure if this param is used by the game.